Update description and begin adding number parsers
The number parsers are not yet complete, but they work for the most
part. A small bug exists that causes the "unsigned_integer" parser to
throw an error on inputs that begin with a non-zero digit. The parser is
short-circuiting due to the `tag("0")` parser, causing the "0" prefix to
be required rather than an optional single digit.