Clean up and clarify crate-level docs and README
The crate-level documentation was messy, verbose and all over the place.
It's safe to say that the README was in the same state of disarray.

Much of this commit is about properly setting up documentation for the
library, re-writing/cleaning up existing docs and making things more
succinct.

The change log has been updated to include the changes thus far. All
contributions, no matter big or small, must be reflected in the
changelog.