Add JSON grammar file, logos and update README
The JSON grammar definition file was added so there is an easily
readable syntax to use when creating parsers, etc. The README was
expanded to expound upon the design decisions made.

Proofreading revealed a few errors and typos within the documentation so
those were fixed. The 'attributions' section was removed but this is
most likely temporary until we can find a better-suited place for it.