Finish float/integer migration to number
Support for JSON's number data type is represented as a Rust f64, or
64-bit floating point type. This will likely change in the future, but
it is too complex of an issue with too many edge cases to justify
spending time on perfecting.

After an initial MVP release, support for more refined Rust types will
be considered. For now, using a unified 'number' type, represented as an
f64, is compliant with the official JSON specification and allows the
project to progress.

The linker config was updated locally  to use the faster mold linker for
Linux targets and zld linker for macOS targets, respectively. The CI
pipeline will be updated to reflect this change, ensuring that this does
not break the build on non-local machines.