Fixed test coverage, docs and specialize helpers
The helper functions for trimming whitespace were monomorphized as
much as possible since we only need to support a specific type of input.
If the input type changes in the future, we can either use more generic
code or simply change the signatures to match the types.

Test coverage should be back to full and all missing documentation was
added.